Move extensions to Funcky.Extensions namespace
